==> QuantSeq/10.GOGeneListProduction.R <==
library(tidyverse)

d14Unsampled <- read_csv("DESeqD14RNASeqUnsampled.csv")
KOUnsampled <- read_csv("DESeqKORNASeqUnsampled.csv")
d14Quant <- read_csv("DESeqD14QuantSeq.csv")
KOQuant <- read_csv("DESeqKOQuantSeq.csv")

d14QuantAdjSig <- filter(d14Quant, padj < 0.05) %>% select(EnsemblID)
d14QuantAdjSig <- as.list(d14QuantAdjSig$EnsemblID)
d14QuantSig <- filter(d14Quant, pvalue < 0.05) %>% select(EnsemblID)
d14QuantSig <- as.list(d14QuantSig$EnsemblID)
KOQuantAdjSig <- filter(KOQuant, padj < 0.05) %>% select(EnsemblID)
KOQuantAdjSig <- as.list(KOQuantAdjSig$EnsemblID)
KOQuantSig <- filter(KOQuant, pvalue < 0.05) %>% select(EnsemblID) 
KOQuantSig <- as.list(KOQuantSig$EnsemblID)

d14UnsampledAdjSig <- filter(d14Unsampled, padj < 0.05) %>% select(EnsemblID) 
d14UnsampledAdjSig <- as.list(d14UnsampledAdjSig$EnsemblID)
d14UnsampledSig <- filter(d14Unsampled, pvalue < 0.05) %>% select(EnsemblID)
d14UnsampledSig <- as.list(d14UnsampledSig$EnsemblID)
KOUnsampledAdjSig <- filter(KOUnsampled, padj < 0.05) %>% select(EnsemblID) 
KOUnsampledAdjSig <- as.list(KOUnsampledAdjSig$EnsemblID)
KOUnsampledSig <- filter(KOUnsampled, pvalue < 0.05) %>% select(EnsemblID) 
KOUnsampledSig <- as.list(KOUnsampledSig$EnsemblID)

write.table(d14QuantAdjSig, "d14QuantAdjSig.txt", quote = FALSE, col.names = FALSE, row.names = FALSE, sep = "\n")
write.table(d14QuantSig, "d14QuantSig.txt", quote = FALSE, col.names = FALSE, row.names = FALSE, sep = "\n")
write.table(KOQuantAdjSig, "KOQuantAdjSig.txt", quote = FALSE, col.names = FALSE, row.names = FALSE, sep = "\n")
write.table(KOQuantSig, "KOQuantSig.txt", quote = FALSE, col.names = FALSE, row.names = FALSE, sep = "\n")
write.table(d14UnsampledAdjSig, "d14UnsampledAdjSig.txt", quote = FALSE, col.names = FALSE, row.names = FALSE, sep = "\n")
write.table(d14UnsampledSig, "d14UnsampledSig.txt", quote = FALSE, col.names = FALSE, row.names = FALSE, sep = "\n")
write.table(KOUnsampledAdjSig, "KOUnsampledAdjSig.txt", quote = FALSE, col.names = FALSE, row.names = FALSE, sep = "\n")
write.table(KOUnsampledSig, "KOUnsampledSig.txt", quote = FALSE, col.names = FALSE, row.names = FALSE, sep = "\n")
